requirements.txt
anthropic==0.43.0
duckduckgo-search==7.3.2
numpy==2.1.3
sentence-transformers==3.3.1
mcp==1.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
